+++ soroban-spec/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod codegen;
-pub mod parse;
-pub mod wasm;
+pub mod read;
